+++ src/main/java/net/larsmans/infinitybuttons/InfinityButtonsInit.java
@@ -7,7 +7,7 @@ import net.larsmans.infinitybuttons.sounds.ModSounds;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class InfinityButtons implements ModInitializer {
+public class InfinityButtonsInit implements ModInitializer {
 	public static final String MOD_ID = "infinitybuttons";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 	public static final net.larsmans.infinitybuttons.InfinityButtonsConfig CONFIG = net.larsmans.infinitybuttons.InfinityButtonsConfig.createAndLoad();