+++ src/main/java/net/larsmans/infinitybuttons/block/custom/EmeraldButton.java
@@ -1,7 +1,7 @@
 package net.larsmans.infinitybuttons.block.custom;
 
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
-import net.larsmans.infinitybuttons.InfinityButtons;
+import net.larsmans.infinitybuttons.InfinityButtonsInit;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.item.ItemStack;
@@ -31,7 +31,7 @@ public class EmeraldButton extends Button{
 
     @Override
     public void appendTooltip(ItemStack stack, @Nullable BlockView world, List<Text> tooltip, TooltipContext options) {
-        if (InfinityButtons.CONFIG.tooltips()) {
+        if (InfinityButtonsInit.CONFIG.tooltips()) {
             if (Screen.hasShiftDown()) {
                 tooltip.add(Text.translatable("infinitybuttons.tooltip.emerald_button").formatted(Formatting.GRAY));
             } else {