+++ src/main/java/net/larsmans/infinitybuttons/mixin/ExampleMixin.java
@@ -1,6 +1,6 @@
 package net.larsmans.infinitybuttons.mixin;
 
-import net.larsmans.infinitybuttons.InfinityButtons;
+import net.larsmans.infinitybuttons.InfinityButtonsInit;
 import net.minecraft.client.gui.screen.TitleScreen;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -11,6 +11,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class ExampleMixin {
 	@Inject(at = @At("HEAD"), method = "init()V")
 	private void init(CallbackInfo info) {
-		InfinityButtons.LOGGER.info("This line is printed by an example mod mixin!");
+		InfinityButtonsInit.LOGGER.info("This line is printed by an example mod mixin!");
 	}
 }