+++ src/main/java/net/larsmans/infinitybuttons/sounds/ModSounds.java
@@ -1,6 +1,6 @@
 package net.larsmans.infinitybuttons.sounds;
 
-import net.larsmans.infinitybuttons.InfinityButtons;
+import net.larsmans.infinitybuttons.InfinityButtonsInit;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
@@ -9,11 +9,11 @@ public class ModSounds {
     public static SoundEvent ALARM = registerSoundEvent("alarm");
 
     private static SoundEvent registerSoundEvent(String name) {
-        Identifier id = new Identifier(InfinityButtons.MOD_ID, name);
+        Identifier id = new Identifier(InfinityButtonsInit.MOD_ID, name);
         return Registry.register(Registry.SOUND_EVENT, id, new SoundEvent(id));
     }
 
     public static void registerSounds() {
-        System.out.println("Registering ModSounds for " + InfinityButtons.MOD_ID);
+        System.out.println("Registering ModSounds for " + InfinityButtonsInit.MOD_ID);
     }
 }