+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Added Iron and Gold buttons, which can only be pressed by shooting an arrow at it
 - Added tooltips to the iron, gold, emerald and sticky copper buttons
-- Fixed some sound volumes being too low
+- Increased the volume of some sounds
 
 ## 1.3.1